V3 compilation errors: Fix spd_read_byte prototype.
Acked-by: Stefan Reinauer <stepan@coresystems.de>



git-svn-id: svn://coreboot.org/repository/coreboot-v3@567 f3766cd6-281f-0410-b1cd-43a5c92072e9

+++ southbridge/amd/cs5536/smbus_initram.c
@@ -343,7 +343,7 @@ int smbus_read_byte(u16 device, u8 address)
  * @param address The address.
  * @return The data from the SMBus packet area or an error of 0xff (i.e. -1).
  */
-int spd_read_byte(u16 device, u8 address)
+u8 spd_read_byte(u16 device, u8 address)
 {
 	return smbus_read_byte(device, address);
 }
